+++ crates/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `moc-cli` Change Log
 
+## 0.1.1-alpha
+
+Released 2022-10-10.
+
+### Bug correction
+
+* Fix error on "boolean" datatype FIELDs
+* Fix "boolean" datatype parsing in tabledata
+
 
 ## 0.1.0-alpha